more refactoring around margins and distances

+++ main/src/com/tramchester/geo/GridPositions.java
@@ -4,19 +4,17 @@ public class GridPositions {
 
     public static boolean withinDistEasting(final GridPosition gridPositionA, final GridPosition gridPositionB, final MarginInMeters rangeInMeters) {
         return getDistEasting(gridPositionA, gridPositionB) <= rangeInMeters.get();
-        //return rangeInMeters.get() >= getDistEasting(gridPositionA, gridPositionB) ;
     }
 
     public static boolean withinDistNorthing(final GridPosition gridPositionA, final GridPosition gridPositionB, final MarginInMeters rangeInMeters) {
         return getDistNorthing(gridPositionA, gridPositionB) <= rangeInMeters.get();
-        //return rangeInMeters.get() >= getDistNorthing(gridPositionA, gridPositionB);
     }
 
-    private static long getDistNorthing(final GridPosition gridPositionA, final GridPosition gridPositionB) {
+    private static int getDistNorthing(final GridPosition gridPositionA, final GridPosition gridPositionB) {
         return Math.abs(gridPositionA.getNorthings() - gridPositionB.getNorthings());
     }
 
-    private static long getDistEasting(final GridPosition gridPositionA, final GridPosition gridPositionB) {
+    private static int getDistEasting(final GridPosition gridPositionA, final GridPosition gridPositionB) {
         return Math.abs(gridPositionA.getEastings() - gridPositionB.getEastings());
     }
